requirements.txt
absl-py==2.3.1
acres==0.5.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
astunparse==1.6.3
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.10.5
charset-normalizer==3.4.3
ci-info==0.3.0
click==8.3.0
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
dm-tree==0.1.9
etelemetry==0.3.1
exceptiongroup==1.3.1
fastapi==0.135.2
filelock==3.20.0
Flask==3.1.2
flatbuffers==25.9.23
fonttools==4.61.1
fsspec==2025.10.0
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.41.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
httplib2==0.31.0
huggingface-hub==0.36.0
idna==3.10
importlib_resources==6.5.2
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keras==3.0.5
kiwisolver==1.4.9
lazy_loader==0.4
learn==1.0.0
libclang==18.1.1
looseversion==1.3.0
lxml==6.0.2
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml-dtypes==0.3.2
mne==1.11.0
mpmath==1.3.0
namex==0.1.0
narwhals==2.16.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==2.2.6
oauthlib==3.3.1
opencv-python==4.13.0.92
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.3
pathlib==1.0.1
pillow==12.0.0
platformdirs==4.5.1
pooch==1.8.2
protobuf==4.25.8
prov==2.1.1
puremagic==1.30
pyarrow==23.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pydot==4.0.1
Pygments==2.19.2
PyMuPDF==1.26.5
pyparsing==3.2.5
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-multipart==0.0.22
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.3
rdflib==7.4.0
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.6.2
sci-kit==0.1.0
scikit-learn==1.7.2
scipy==1.15.3
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
starlette==1.0.0
streamlit==1.54.0
sympy==1.14.0
tenacity==9.1.4
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-estimator==2.13.0
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.0+cpu
torchaudio==2.9.0+cpu
torchvision==0.24.0+cpu
tornado==6.5.4
tqdm==4.67.1
traits==7.0.2
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.42.0
watchdog==6.0.0
Werkzeug==3.1.3
wrapt==1.17.3
